feat: refactoring to helm chart management for better context understanding
