+++ core/modules/helm_management/helm_values_chart_management/helm_deploy_env_var_filter.py
@@ -2,7 +2,7 @@
 
 
 import re
-from copy import deepcopy
+
 
 # Domain
 from core.modules.data_management import BaseDataHandler
